+++ app/controllers/application_controller.rb
@@ -3,6 +3,4 @@ class ApplicationController < ActionController::Base
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
-
-  include ControllerHelpers
 end
