+++ app/controllers/welcome_controller.rb
@@ -23,9 +23,6 @@ class WelcomeController < ApplicationController
         end
       end
     end
-
-    @data = ControllerHelpers.for_table @metrics
-    @widest = ControllerHelpers.widest @data
   end
 
   private
