+++ app/views/longterm/index.html.erb
@@ -1,3 +1,4 @@
+<% require "pry" ; binding.pry %>
 <% @metrics.each_with_index do |metric, index| %>
   <div class='row'>
     <div class='col-xs-8'>
